Magisk Delta only

+++ native/src/init/rootdir.cpp
@@ -61,13 +61,13 @@ static void patch_init_rc(const char *src, const char *dest, const char *tmp_dir
     clone_attr(src, dest);
 }
 
-static void load_overlay_rc(const char *overlay) {
+static void load_overlay_rc(const char *overlay, bool should_unlink = true) {
     auto dir = open_dir(overlay);
     if (!dir) return;
 
     int dfd = dirfd(dir.get());
     // Do not allow overwrite init.rc
-    unlinkat(dfd, "init.rc", 0);
+    if (should_unlink) unlinkat(dfd, "init.rc", 0);
 
     // '/' + name + '\0'
     char buf[NAME_MAX + 2];
@@ -84,7 +84,7 @@ static void load_overlay_rc(const char *overlay) {
             int rc = xopenat(dfd, entry->d_name, O_RDONLY | O_CLOEXEC);
             rc_list.push_back(full_read(rc));
             close(rc);
-            unlinkat(dfd, entry->d_name, 0);
+            if (should_unlink) unlinkat(dfd, entry->d_name, 0);
         }
     }
 }
@@ -192,7 +192,7 @@ void MagiskInit::patch_ro_root() {
     if (access("/sbin", F_OK) == 0) {
         tmp_dir = "/sbin";
     } else {
-        char buf[8];
+        char buf[16];
         gen_rand_str(buf, sizeof(buf));
         tmp_dir = "/dev/"s + buf;
         xmkdir(tmp_dir.data(), 0);
@@ -228,6 +228,7 @@ void MagiskInit::patch_ro_root() {
 #endif
 
     load_overlay_rc(ROOTOVL);
+    load_overlay_rc(INTLROOT "/early-mount.d/initrc.d", false);
     if (access(ROOTOVL "/sbin", F_OK) == 0) {
         // Move files in overlay.d/sbin into tmp_dir
         mv_path(ROOTOVL "/sbin", ".");
@@ -284,10 +285,6 @@ void MagiskInit::patch_rw_root() {
     }
     rm_rf("/.backup");
 
-    // Patch init.rc
-    patch_init_rc("/init.rc", "/init.p.rc", "/sbin");
-    rename("/init.p.rc", "/init.rc");
-
     bool treble;
     {
         auto init = mmap_data("/init");
@@ -296,6 +293,14 @@ void MagiskInit::patch_rw_root() {
 
     xmkdir(PRE_TMPDIR, 0);
     setup_tmp(PRE_TMPDIR);
+
+    // Handle custom rc script
+    load_overlay_rc(PRE_TMPDIR "/" MIRRDIR "/early-mount/initrc.d", false);
+
+    // Patch init.rc
+    patch_init_rc("/init.rc", "/init.p.rc", "/sbin");
+    rename("/init.p.rc", "/init.rc");
+
     chdir(PRE_TMPDIR);
 
     // Extract magisk
